Split helper tests
